move startup instrumentation to module
