+++ instrumentation/startup/src/main/java/io/opentelemetry/android/instrumentation/startup/AppStartupTimer.java
@@ -12,9 +12,8 @@ import android.os.Handler;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import io.opentelemetry.android.DefaultingActivityLifecycleCallbacks;
+import io.opentelemetry.android.instrumentation.common.DefaultingActivityLifecycleCallbacks;
 import io.opentelemetry.android.common.RumConstants;
-import io.opentelemetry.android.util.AnchoredClock;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.Tracer;
 import io.opentelemetry.sdk.common.Clock;
