+++ settings.gradle.kts
@@ -6,3 +6,4 @@ include(":instrumentation:common-api")
 include(":instrumentation:okhttp:okhttp-3.0:agent")
 include(":instrumentation:okhttp:okhttp-3.0:library")
 include(":instrumentation:okhttp:okhttp-3.0:testing")
+include(":instrumentation:startup")
